fix(tinygo): add sql lite build profile

+++ sql/types/number.go
@@ -1098,22 +1098,22 @@ func convertToUint64(t NumberTypeImpl_, v any, round Round) (uint64, sql.Convert
 		return uint64(v), sql.InRange, nil
 	case int8:
 		if v < 0 {
-			return uint64(math.MaxUint64 - uint(-v-1)), sql.Underflow, nil
+			return uint64(math.MaxUint64 - uint64(-v-1)), sql.Underflow, nil
 		}
 		return uint64(v), sql.InRange, nil
 	case int16:
 		if v < 0 {
-			return uint64(math.MaxUint64 - uint(-v-1)), sql.Underflow, nil
+			return uint64(math.MaxUint64 - uint64(-v-1)), sql.Underflow, nil
 		}
 		return uint64(v), sql.InRange, nil
 	case int32:
 		if v < 0 {
-			return uint64(math.MaxUint64 - uint(-v-1)), sql.Underflow, nil
+			return uint64(math.MaxUint64 - uint64(-v-1)), sql.Underflow, nil
 		}
 		return uint64(v), sql.InRange, nil
 	case int64:
 		if v < 0 {
-			return uint64(math.MaxUint64 - uint(-v-1)), sql.Underflow, nil
+			return uint64(math.MaxUint64 - uint64(-v-1)), sql.Underflow, nil
 		}
 		return uint64(v), sql.InRange, nil
 	case uint:
@@ -1131,7 +1131,7 @@ func convertToUint64(t NumberTypeImpl_, v any, round Round) (uint64, sql.Convert
 			return math.MaxUint64, sql.Overflow, nil
 		}
 		if v < 0 {
-			return uint64(math.MaxUint64 - uint(-v-1)), sql.Underflow, nil
+			return uint64(math.MaxUint64 - uint64(-v-1)), sql.Underflow, nil
 		}
 		return uint64(math.Round(float64(v))), sql.InRange, nil
 	case float64:
@@ -1142,7 +1142,7 @@ func convertToUint64(t NumberTypeImpl_, v any, round Round) (uint64, sql.Convert
 			return math.MaxUint64, sql.Overflow, nil
 		}
 		if v < 0 {
-			return uint64(math.MaxUint64 - uint(-v-1)), sql.Underflow, nil
+			return uint64(math.MaxUint64 - uint64(-v-1)), sql.Underflow, nil
 		}
 		return uint64(math.Round(v)), sql.InRange, nil
 	case *apd.Decimal:
